refactor of names for compliance
