+++ drake/solvers/CMakeLists.txt
@@ -58,9 +58,9 @@ else()
   list(APPEND optimization_files NoMosek.cpp)
 endif()
 if(gurobi_FOUND)
-  list(APPEND optimization_files GurobiSolver.cc)
+  list(APPEND optimization_files gurobi_solver.cc)
 else()
-  list(APPEND optimization_files NoGurobi.cc)
+  list(APPEND optimization_files no_gurobi.cc)
 endif()
 
 add_library_with_exports(LIB_NAME drakeOptimization SOURCE_FILES ${optimization_files})
