+++ drake/solvers/no_gurobi.cc
